refactor: Extracted function
