+++ src/utils/sort-invocations/index.ts
@@ -1,3 +1,4 @@
+export * from './list-splattributes-last.js';
 export * from './sort-attributes.js';
 export * from './sort-hash.js';
 export * from './sort-modifiers.js';
